split up command
